+++ fhir_auth/lib/fhir_client/export.dart
@@ -1,3 +1,4 @@
+export 'epc_fhir_client.dart';
 export 'fhir_client.dart';
 export 'gcp_fhir_client.dart';
 export 'secure_fhir_client.dart';