+++ fhir_auth/lib/fhir_client/smart_fhir_client.dart
@@ -12,6 +12,7 @@ import 'package:oauth2/oauth2.dart';
 
 // Project imports:
 import 'authenticate/base_authentication.dart';
+import 'authenticate/smart_authorization_code_grant.dart';
 import 'secure_fhir_client.dart';
 
 import 'authenticate/authenticate.dart'
@@ -22,23 +23,57 @@ import 'authenticate/authenticate.dart'
 
 class SmartFhirClient extends SecureFhirClient {
   SmartFhirClient({
-    required FhirUri fhirUri,
-    required String clientId,
-    required FhirUri redirectUri,
+    required super.fhirUri,
+    required super.clientId,
+    required super.redirectUri,
     this.customUriScheme,
-    List<String>? scopes,
+    super.scopes,
     this.authorizeUrl,
     this.tokenUrl,
+    super.launch,
+    super.secret,
+  });
+
+  factory SmartFhirClient.fromLaunchParameters(
+    Uri base,
+    Map<String, String> queryParameters, {
+    FhirUri? fhirUri,
+    String? clientId,
+    FhirUri? redirectUri,
+    String? customUriScheme,
+    List<String>? scopes,
+    String? authorizeUrl,
+    String? tokenUrl,
     String? launch,
     String? secret,
-  }) : super(
-          fhirUri: fhirUri,
-          clientId: clientId,
-          redirectUri: redirectUri,
-          scopes: scopes,
-          launch: launch,
-          secret: secret,
-        );
+    String redirectPath = '/redirect.html',
+  }) {
+    fhirUri ??= queryParameters['iss'] == null
+        ? throw Exception('no fhirUri was passed for SMART launch')
+        : FhirUri(queryParameters['iss']);
+    launch ??= queryParameters['launch'];
+    clientId ??= queryParameters['clientId'];
+    redirectUri ??= FhirUri(Uri(
+      host: base.host,
+      scheme: base.scheme,
+      port: base.port,
+      path: redirectPath,
+    ));
+    if (clientId == null) {
+      throw Exception('no clientId was passed for SMART launch');
+    }
+    return SmartFhirClient(
+      fhirUri: fhirUri,
+      clientId: clientId,
+      redirectUri: redirectUri,
+      customUriScheme: customUriScheme,
+      scopes: scopes,
+      authorizeUrl: authorizeUrl == null ? null : FhirUri(authorizeUrl),
+      tokenUrl: tokenUrl == null ? null : FhirUri(tokenUrl),
+      launch: launch,
+      secret: secret,
+    );
+  }
 
   String? customUriScheme;
   FhirUri? authorizeUrl;
@@ -46,50 +81,115 @@ class SmartFhirClient extends SecureFhirClient {
   Uri? responseUrl;
   BaseAuthentication authClient = createAuthentication();
   Client? client;
+
+  /// https://build.fhir.org/ig/HL7/smart-app-launch/scopes-and-launch-context.html#launch-context-arrives-with-your-access_token
+  /// Once an app is authorized, the token response will include any context
+  /// data the app requested and any (potentially) unsolicited context data the
+  /// EHR may decide to communicate. For example, EHRs may use launch context to
+  /// communicate UX and UI expectations to the app (see need_patient_banner below).
+  ///
+  /// String value with a patient id, indicating that the app was launched in
+  /// the context of FHIR Patient 123. If the app has any patient-level scopes,
+  /// they will be scoped to Patient 123.
   String? patientId;
+
+  /// String value with an encounter id, indicating that the app was launched
+  /// in the context of FHIR Encounter 123.
   String? encounterId;
 
+  /// Boolean value indicating whether the app was launched in a UX context
+  /// where a patient banner is required (when true) or not required (when
+  /// false). An app receiving a value of false should not take up screen real
+  /// estate displaying a patient banner.
+
+  bool? needPatientBanner;
+
+  /// String URL where the EHR’s style parameters can be retrieved (for apps
+  /// that support styling)
+
+  String? smartStyleUrl;
+
+  /// Array of relative resource References to any resource type other than
+  /// “Patient” or “Encounter”. It is not prohibited to have more than one
+  /// Reference to a given type of resource.
+
+  List<String>? fhirContext;
+
+  ///String value describing the intent of the application launch (see notes below)
+
+  String? intent;
+
+  /// String conveying an opaque identifier for the healthcare organization
+  /// that is launching the app. This parameter is intended primarily to support
+  /// EHR Launch scenarios.
+  String? tenant;
+
   Future<void> login() async {
+    /// if no authorizeUrl or tokenUrl, go find them
     if (authorizeUrl == null || tokenUrl == null) {
       final capabilityStatement = await _getCapabilityStatement();
       authorizeUrl = _getUri(capabilityStatement, 'authorize');
       tokenUrl = _getUri(capabilityStatement, 'token');
     }
 
-    final grant = AuthorizationCodeGrant(
+    /// Create a new code grant (had to extend it, because FHIR uses
+    /// non-standard parameters)
+    final grant = SmartAuthorizationCodeGrant(
       clientId!,
       authorizeUrl!.value!,
       tokenUrl!.value!,
       secret: secret,
     );
+
+    /// Get the authorizationUrl
     var authorizationUrl = grant.getAuthorizationUrl(
       redirectUri!.value!,
       scopes: scopes,
     );
+
+    /// Create a map from the parameters
     final params = Map.of(authorizationUrl.queryParameters);
+
+    /// Add the aud in
     params['aud'] = '$fhirUri';
+
+    /// Replace those params as part of the authorizationUrl
     authorizationUrl = authorizationUrl.replace(queryParameters: params);
+
     try {
+      /// Attempt to authenticate
       final returnValue = await authClient.authenticate(
         authorizationUrl: authorizationUrl,
         redirectUri: redirectUri!,
       );
+
+      /// Use the returnValue's parameters to handle Authorization
+      /// NOTE: as part of the authorization, we will also get other fhir
+      /// parameters that may be passed that are not officially part of an
+      /// oauth2 workflow
       client = await grant
           .handleAuthorizationResponse(Uri.parse(returnValue).queryParameters);
-      print(client!.credentials.toJson());
-      // print((jsonDecode(client!.credentials.toJson()) as Map).keys);
-      // print(client!.credentials.scopes);
-      // print(client!.credentials.accessToken);
-      // print(client!.credentials.refreshToken);
 
-      // print(client?.credentials.accessToken);
-      // print(client?.credentials.)
-      // print(jsonDecode(jsonEncode(client?.credentials.toJson())).keys);
-      // print(client?.identifier);
-      // print(client?.toString());
+      /// assign context values if available
+      patientId = grant.fhirParameters['patient'];
+      encounterId = grant.fhirParameters['encounter'];
+      needPatientBanner = grant.fhirParameters['need_patient_banner'] == null
+          ? null
+          : grant.fhirParameters['need_patient_banner'].toString() == 'true'
+              ? true
+              : grant.fhirParameters['need_patient_banner'].toString() ==
+                      'false'
+                  ? false
+                  : null;
+      smartStyleUrl = grant.fhirParameters['smart_style_url'];
+      fhirContext = grant.fhirParameters['fhirContext'];
+      intent = grant.fhirParameters['intent'];
+      tenant = grant.fhirParameters['tenant'];
     } catch (e, stack) {
       log('Exception: $e');
       log('Stack: $stack');
+      print('Exception: $e');
+      print('Stack: $stack');
     }
   }