+++ fhir_auth/lib/scopes/dstu2/clinical_scope.dart
@@ -9,6 +9,41 @@ enum Interaction {
   read,
   write,
   any,
+
+  /// 5
+  cruds,
+
+  /// 4
+  ruds,
+  cuds,
+  crds,
+  crus,
+  crud,
+
+  /// 3
+  uds,
+  rds,
+  rud,
+  cds,
+  cud,
+  rus,
+  cus,
+  cru,
+
+  /// 2
+  cr,
+  cu,
+  cd,
+  cs,
+  ru,
+  rd,
+  rs,
+  ud,
+  us,
+  ds,
+
+  /// 1
+
   c,
   r,
   u,
@@ -20,14 +55,21 @@ enum Interaction {
 enum Role {
   patient,
   user,
+  system,
 }
 
 /// Together these 3 items make up a FHIR Clinical scope
 /// http://www.hl7.org/fhir/smart-app-launch/scopes-and-launch-context.html#clinical-scope-syntax
 class ClinicalScope {
-  ClinicalScope(this.role, this.resourceType, this.interaction);
+  ClinicalScope({
+    required this.role,
+    this.resourceType,
+    bool? allTypes,
+    required this.interaction,
+  }) : allTypes = resourceType == null ? true : allTypes ?? false;
 
   final Role role;
-  final Dstu2ResourceType resourceType;
+  final Dstu2ResourceType? resourceType;
+  final bool allTypes;
   final Interaction interaction;
 }